Refactor dataset_repo
* Marks SQLRepository as abstract

* Drops 'save' method on all implementations of SQLRepository

* Enforces SQLRepository implementations to define table_name

* Rename dataset_repository to dataframe_repository
